+++ src/scanoss/inspection/undeclared_component.py
@@ -23,7 +23,7 @@
 """
 import json
 from typing import Dict, Any
-from scanoss.inspection.policy_check import PolicyCheck, PolicyStatus
+from .policy_check import PolicyCheck, PolicyStatus
 
 class UndeclaredComponent(PolicyCheck):
     """
@@ -32,7 +32,7 @@ class UndeclaredComponent(PolicyCheck):
     """
 
     def __init__(self, debug: bool = False, trace: bool = True, quiet: bool = False, filepath: str = None,
-                 format_type: str = 'json', status: str = None, output: str = None):
+                 format_type: str = 'json', status: str = None, output: str = None, sbom_format: str = 'settings'):
         """
         Initialize the UndeclaredComponent class.
 
@@ -43,6 +43,7 @@ class UndeclaredComponent(PolicyCheck):
         :param format_type: Output format ('json' or 'md')
         :param status: Path to save status output
         :param output: Path to save detailed output
+        :param sbom_format: Sbom format for status output (default 'settings')
         """
         super().__init__(debug, trace, quiet, filepath, format_type, status, output,
                          name='Undeclared Components Policy')
@@ -50,6 +51,7 @@ class UndeclaredComponent(PolicyCheck):
         self.format = format
         self.output = output
         self.status = status
+        self.sbom_format = sbom_format
 
     def _get_undeclared_component(self, components: list)-> list or None:
         """
@@ -59,7 +61,7 @@ class UndeclaredComponent(PolicyCheck):
            :return: List of undeclared components
         """
         if components is None:
-            self.print_stderr(f'WARNING: No components provided!')
+            self.print_debug(f'WARNING: No components provided!')
             return None
         undeclared_components = []
         for component in components:
@@ -78,9 +80,14 @@ class UndeclaredComponent(PolicyCheck):
         """
         summary = f'{len(components)} undeclared component(s) were found.\n'
         if len(components) > 0:
+            if self.sbom_format == 'settings':
+                summary += (f'Add the following snippet into your `scanoss.json` file\n'
+                            f'\n```json\n{json.dumps(self._generate_scanoss_file(components), indent=2)}\n```\n')
+                return summary
+
             summary += (f'Add the following snippet into your `sbom.json` file\n'
                         f'\n```json\n{json.dumps(self._generate_sbom_file(components), indent=2)}\n```\n')
-        return summary
+            return summary
 
     def _json(self, components: list) -> Dict[str, Any]:
         """
@@ -115,20 +122,49 @@ class UndeclaredComponent(PolicyCheck):
             'summary': self._get_summary(components),
         }
 
-    def _generate_sbom_file(self, components: list) -> list:
+    def _get_unique_components(self, components: list) -> list:
+        """
+         Generate a list of unique components.
+
+         :param components: List of undeclared components
+         :return: list of unique components
+         """
+        unique_components = {}
+        if components is None:
+            self.print_stderr(f'WARNING: No components provided!')
+            return []
+
+        for component in components:
+                unique_components[component['purl']] = {'purl': component['purl']}
+        return list(unique_components.values())
+
+    def _generate_scanoss_file(self, components: list) -> dict:
+        """
+         Generate a list of PURLs for the scanoss.json file.
+
+         :param components: List of undeclared components
+         :return: scanoss.json Dictionary
+         """
+        scanoss_settings = {
+            'bom':{
+                'include': self._get_unique_components(components),
+            }
+        }
+
+        return scanoss_settings
+
+    def _generate_sbom_file(self, components: list) -> dict:
         """
          Generate a list of PURLs for the SBOM file.
 
          :param components: List of undeclared components
-         :return: List of dictionaries containing PURLs
+         :return: SBOM Dictionary with components
          """
-        sbom = {}
-        if components is None:
-            self.print_stderr(f'WARNING: No components provided!')
-        else:
-            for component in components:
-                sbom[component['purl']] = { 'purl': component['purl'] }
-        return list(sbom.values())
+        sbom = {
+            'components': self._get_unique_components(components),
+        }
+
+        return sbom
 
     def run(self):
         """